+++ lib/DOM/actions/toggle_attribute.ex
@@ -1,19 +1,19 @@
 defmodule Tamnoon.DOM.Actions.ToggleAttribute do
   @moduledoc """
   An action that will toggle `:attribute` attribute of the `:target` node. An additional
-  argument `:force` can also be provided in order to force the value as `true` or `false`.
+  argument `:force_to` can also be provided in order to force the value as `true` or `false`.
   """
   alias Tamnoon.DOM
   import DOM
   use DOM.JsonEncoder, type: :action
 
   @enforce_keys [:target, :attribute]
-  defstruct [:target, :attribute, :force]
+  defstruct [:target, :attribute, :force_to]
 
   @type t :: %__MODULE__{
           target: DOM.Node.t(),
           attribute: String.t(),
-          force: boolean() | nil
+          force_to: boolean() | nil
         }
 
         @doc """
@@ -25,8 +25,8 @@ defmodule Tamnoon.DOM.Actions.ToggleAttribute do
 
   def is_valid?(%DOM.Actions.ToggleAttribute{}), do: true
 
-  def is_valid?(%{target: %DOM.Node{}, attribute: attribute, force: force})
-      when is_binary(attribute) and is_boolean(force),
+  def is_valid?(%{target: %DOM.Node{}, attribute: attribute, force_to: force_to})
+      when is_binary(attribute) and is_boolean(force_to),
       do: true
 
   def is_valid?(%{target: %DOM.Node{}, attribute: attribute})
@@ -42,7 +42,7 @@ defmodule Tamnoon.DOM.Actions.ToggleAttribute do
 
   * `:attribute`: a `t:String.t/0`.
 
-  * `:force`: a boolean (optional).
+  * `:force_to`: a boolean (optional).
   """
   @spec new!(action_args :: term()) :: %DOM.Actions.ToggleAttribute{}
   def new!(%DOM.Actions.ToggleAttribute{} = action_args), do: action_args