+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### v1.0.0-rc.2
 
+- Renamed the `:force` attribute in `m:Tamnoon.DOM.Actions.ToggleAttribute` to `:force_to`. 
+
 - Fixed a bug where DOM elements added via actions would not be able to trigger methods.
 
 - Fixed a missing app name injection in the success message for the `m:Mix.Tasks.Tamnoon.OverrideRoot` task.